Add support for CKTransactionalComponentDataSourceChangeset verification
